переделал сервлет под тег либы, начал реализовывать сервис сделал репозиторий спринг дата сделал пул коннектов хикари

+++ isfmo-logic/src/main/java/org/bitbucket/eniqen/dao/BaseDao.java
@@ -15,7 +15,7 @@ public interface BaseDao<T extends Serializable> {
 
     List<T> getAll();
 
-    T getById(Integer id);
+    T getById(Long id);
 
     void add(T model);
